ai-service/requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
python-dotenv
scikit-learn
requests
tldextract
python-whois
